# cmd/nube/openapi.yaml
openapi: 3.0.3
info:
  title: Nube Server API
  version: 0.2.0
  description: >
    Multi-tenant central server for NubeIO — workspace/user management,
    app catalog, install flow, per-user MCP tool scoping, agent execution,
    app store with reviews and sharing.
servers:
  - url: http://localhost:8090
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      summary: Get user by ID (admin)
      x-cli:
        command: users get
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: User ID
      responses:
        "200":
          description: OK

    delete:
      operationId: deleteUser
      summary: Delete user (admin)
      x-cli:
        command: users delete
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: User ID
      responses:
        "200":
          description: Deleted
...
  /apps:
    get:
      operationId: listApps
      summary: List available apps from catalog
      x-cli:
        command: apps list
      responses:
        "200":
          description: OK

    post:
      operationId: createApp
      summary: Create a new app on disk (admin)
      x-cli:
        command: apps create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string, description: "App slug (lowercase, hyphens, 3-50 chars)" }
                description: { type: string, description: App description }
                author:      { type: string, description: Author name }
                version:     { type: string, description: "Semver (default: 1.0.0)" }
                tags:        { type: array, items: { type: string }, description: Search tags }
      responses:
        "201":
          description: Created
...
  /api/agents/jobs:
    post:
      operationId: submitJob
      summary: Submit an async AI job (returns immediately with job ID)
      # x-cli: hand-written command in cmd_agents.go (supports streaming poll)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:   { type: string, description: The prompt to send }
                agent:    { type: string, description: "Agent/app name (optional)" }
                provider: { type: string, description: "AI provider (default: claude)" }
                model:    { type: string, description: "Model override" }
      responses:
        "202":
          description: Job submitted
  /api/agents/jobs/{id}:
    get:
      operationId: pollJob
      summary: Poll job status and events
      # x-cli: hand-written command in cmd_agents.go (supports streaming poll)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Job ID
        - name: after
          in: query
          required: false
          schema: { type: integer }
          description: "Only return events with index > after (incremental polling)"
      responses:
        "200":
          description: Job state and events
    delete:
      operationId: cancelJob
      summary: Cancel a running job
      # x-cli: could add hand-written cancel command later
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Job ID
      responses:
        "200":
          description: Job cancelled
...